refactor: client scripts
