+++ esbuild.js
@@ -62,11 +62,11 @@ async function main() {
   // Script for Webview build
   const mediaCtx = await esbuild.context({
     entryPoints: [
-      "media/abcjsScripts.mts",
-      "media/mermaidScripts.mts",
-      "media/vscodeWebviewScrollScripts.mts",
-      "media/webSocketScripts.mts",
-      "media/youtubePlaceholderScripts.mts",
+      "media/client/browser.mts",
+      "media/client/webview.mts",
+      "media/client/shared/abcjsScripts.mts",
+      "media/client/shared/mermaidScripts.mts",
+      "media/client/shared/youtubePlaceholderScripts.mts",
     ],
     bundle: true,
     format: "iife",
