+++ media/client/shared/mermaidScripts.mts
