+++ media/client/shared/youtubePlaceholderScripts.mts
